+++ src/pb-components.js
@@ -59,6 +59,7 @@ import './pb-blacklab-highlight.js';
 import './pb-table-grid.js';
 import './pb-split-list.js';
 import './pb-timeline.js';
+import './pb-combo-box.js';
 
 import '@polymer/iron-icons/editor-icons';
 import '@polymer/iron-icons/social-icons';